copy PMM module from MSP430 DriverLib (MSPWare)
This is not nice. At least we should make this library depend on
mspware. Problem is that moving to MSPWare ideally would involves a
giant re-write of all apps to replace register access with MSPWare API
calls. So, until we decide to do that, we predent that
we don't use MSPWare -- so, we silently sneak in some MSPWare
ware code into libmsp.