uart: fix waiting for last byte to tx
We want TXCP interrupt from last byte, not from any previous byte, so we
have to start waiting for TXCP interrupt only after TX interrupt that
corresponds to BUF clearing up after the last byte. Otherwise, we catch
TXCP from second-to-last byte and prematurely declare the send as
completed.

Also, enqueue the first byte (which clears TXIFG) and
then enable the TXIE. This seems cleaner, but shouldn't
make a difference.